requirements.txt
asgiref==3.7.2
Django==5.0.3
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
PyJWT==1.7.1
sqlparse==0.4.4
django-cors-headers==4.3.1
